feat: forced_theme option (#62)
* option for forcing a light or dark theme

a `force_theme` option is now available in the `[extra]` section of the `config.toml` file. when set to `false`, existing serene behavior is kept: the browser light dark mode is honored and a toggling button is included at the end of the "links" section of the home page

when `force_theme` is set to `"light"` or `"dark"`, the chosen theme is set for all clients, and the toggling button is removed.

## implementation

when `force_theme` is set, the toggling `<button>` element is replaced by an invisible `<span>` that has the same id (`"theme-toggle"`). this `<span>` has an `force_theme` attribute containing the theme from `force_theme` option.

the JS routine `enableThemeToggle()` looks for this `force_theme` attribute. if found, it toggles to the given theme and skips adding the event listeners for theme changes.

* correct syntax for getting forced theme attribute

* documentation for forced_theme

- simplified chinese is via google translate.

* add icons for more social networks

oldies:
 - facebook
 - linkedin
 - instagram
newbs:
 - bluesky
 - mastodon
 - threads

* fix: forece_theme

---------

Co-authored-by: Jianan Sun <isunjn@gmail.com>